bugs fixes +testing of register via postman

+++ src/utils/cloudinary.js
@@ -1,5 +1,8 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs";
+import dotenv from "dotenv";
+
+dotenv.config({ path: "./.env" });
 
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
@@ -9,15 +12,19 @@ cloudinary.config({
 
 const uploadOnCloudniary=async(localFilePath)=>{
     try {
-        if (!localFilePath) {return null;}
+        if (!localFilePath) return null;
         const response=await cloudinary.uploader.upload(localFilePath,{
             resource_type:"auto"
-        })
-        console.log("file has been uploaded successfully",response.url);
+        });
+        // console.log(" File uploaded to Cloudinary:", response.secure_url);
+        fs.unlinkSync(localFilePath);
+       
         return response;
            
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        console.error(" Cloudinary upload error:", error);
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
+
         return null;
     }
 }
